Extract Output to its own component.
